Lots of Stuff
The main thing to pay attention to are the new "ChapterXFCDialogue.asm" files. Making these files is necessary for organizational purposes. Needless to say, all FC dialogue goes in there. Just to be safe, we'll also put vanilla conversations that are now longer in FC in there. For example, the dialogue for Travant and Arion taking to the field in chapter 14.